+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # flatlog Changelog
 
+## 1.0.2 - 2026-07-13
+- Added: accept `v` prefix and prerelease suffix in version headers
+- Added: seed `init` initial release from `package.json` version when valid
+- Added: support `--dry-run` for `next` command
+
 ## 1.0.1 - 2026-06-10
 - Added: `next command` to open a new unreleased
 - Fixed: simplified error and warning messages